functions/notify-customer/requirements.txt
functions-framework==3.*
flask
sendgrid
google-cloud-bigquery